Change timescale base from ticks/long to seconds/double

+++ FerretLib.SFML/ChronoEventArgs.cs
@@ -4,7 +4,7 @@ namespace FerretLib.SFML
 {
     public class ChronoEventArgs : EventArgs
     {
-        public ChronoEventArgs(long monotonic, double delta, double fps)
+        public ChronoEventArgs(double monotonic, double delta, double fps)
         {
             Monotonic = monotonic;
             Delta = delta;
@@ -12,7 +12,7 @@ namespace FerretLib.SFML
             Fps = fps;
         }
 
-        public long Monotonic { get; private set; }
+        public double Monotonic { get; private set; }
         public double Delta { get; private set; }
         public double Fps { get; private set; }
         public DateTime DateTime { get; private set; }
